watchdog: Implement status function in watchdog core
Up to now, the watchdog status function called a driver function,
which was supposed to return the watchdog status. All but one
driver did not implement this function, and the driver implementing
it did not implement it correctly (the function is supposed to
return WDIOF_ flags).

Re-implement the function in the watchdog driver core, effectively
making the callback into the driver code unnecessary. This also
makes the 'ststus' sysfs attribute always available. The status attribute
is now displayed as hex number with 0x prepended to be easier to decode.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>