Input: sirfsoc-onkey - Drop unnecessary cleanup calls and other changes
Calling dev_set_drvdata() or device_init_wakeup() from a
driver's remove function is unnecessary and can be dropped.
Other changes as listed below.

The conversion was done automatically with coccinelle using the
following semantic patches. The semantic patches and the scripts used
to generate this commit log are available at
https://github.com/groeck/coccinelle-patches

- Drop empty remove function
- Drop 'device_init_wakeup();' from remove function

Cc: Barry Song <baohua@kernel.org>
Signed-off-by: Guenter Roeck <linux@roeck-us.net>